+++ mappings/muse.py
@@ -1,0 +1,106 @@
+from mappings.marc import MARCMapping
+
+class MUSEMapping(MARCMapping):
+    def __init__(self, source):
+        super(MUSEMapping, self).__init__(source, {})
+        self.mapping = self.createMapping()
+    
+    def createMapping(self):
+        return {
+            'identifiers': [
+                ('001', '{0}|muse'),
+                ('010', '{z}|lccn'),
+                ('020', '{a}{z}|isbn'),
+                ('022', '{a}|issn'),
+                ('035', '{a}|oclc'),
+                ('040', '{a}|muse')
+            ],
+            'authors': [
+                ('100', '{a} {b} {c} {d}|||true'),
+                ('110', '{a} {b} {c} {n} {d}|||true'),
+            ],
+            'title': [
+                ('245', '{a} {b}'),
+                ('130', '{a}')
+            ],
+            'alternative': [
+                ('240', '{a} {k}') # Uniform Title
+            ],
+            'has_version': ('250', '{a} {b}|'),
+            'publisher': ('264', '{b}||'),
+            'spatial': ('264', '{a}'),
+            'dates': [
+                ('264', '{c}|publication_date')
+            ],
+            'languages': [('008', '||{0}')],
+            'extent': ('300', '{a}{b}{c}'),
+            'table_of_contents': ('505', '{a}'),
+            'abstract': [
+                ('500', '{a}'),
+                ('504', '{a}'),
+                ('520', '{a}'),
+            ],
+            'subjects': [
+                ('600', '{a} {d} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('610', '{a} {d} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('611', '{a} {d} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('630', '{a} {p} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('650', '{a} {b} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('651', '{a} -- {v} -- {x} -- {y} -- {z}|{2}|{0}'),
+                ('656', '{a} -- {v} -- {x} -- {y} -- {z}|{2}|{0}')
+            ],
+            'contributors': [
+                ('260', '{f}|||manufacturer'),
+                ('700', '{a} {b} {c} {d}|||{e}'),
+                ('710', '{a} {b} {c} {d}|||{e}'),
+                ('711', '{a} {e}|||{j}'),
+            ],
+            'is_part_of': ('490', '{a}|{v}|volume'),
+            'has_part': [
+                ('856', '1|{u}|muse|text/html|{{"reader": false, "download": false, "catalog": false}}')
+            ]
+        }
+
+    def applyFormatting(self):
+        self.record.source = 'muse'
+        self.record.source_id = list(self.record.identifiers[0].split('|'))[0]
+
+        # Take first title as they are in order of preference
+        self.record.title = self.record.title[0]
+
+        # Extract language code from 008 fixed data field
+        self.record.languages = [self.extractLanguage(l) for l in self.record.languages]
+
+        # Clean up subjects to remove spots for missing subheadings
+        self.record.subjects = [
+            self.cleanUpSubjectHead(s)
+            for s in self.record.subjects
+        ]
+
+    def cleanUpSubjectHead(self, subject):
+        subjectStr, *subjectMeta = subject.split('|')
+        subjectParts = subjectStr.split('--')
+
+        outParts = []
+
+        for part in subjectParts:
+            cleanPart = part.strip(' .')
+            
+            if cleanPart == '': continue
+
+            outParts.append(cleanPart)
+
+        cleanSubject = ' -- '.join([p for p in outParts])
+
+        return '|'.join([cleanSubject] + subjectMeta)
+
+    def extractLanguage(self, language):
+        _, _, marcData, *_ = language.split('|')
+        return '||{}'.format(marcData[35:38])
+
+    def addHasPartLink(self, url, mediaType, flags):
+        lastItemNo = int(self.record.has_part[0][0])
+
+        self.record.has_part.append(
+            '{}|{}|muse|{}|{}'.format(lastItemNo + 1, url, mediaType, flags)
+        )